requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astroid @ file:///tmp/build/80754af9/astroid_1592495912941/work
certifi==2020.12.5
chardet==4.0.0
Django==3.0.5
djangorestframework==3.12.2
djongo==1.3.3
idna==2.10
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
lazy-object-proxy @ file:///tmp/build/80754af9/lazy-object-proxy_1607707356794/work
mailjet-rest==1.3.4
mccabe==0.6.1
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1608147593446/work
pylint @ file:///tmp/build/80754af9/pylint_1598623985952/work
pymongo==3.11.2
python-http-client==3.3.1
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
requests==2.25.1
sendgrid==6.5.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
sqlparse==0.2.4
starkbank-ecdsa==1.1.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
urllib3==1.26.2
wrapt==1.11.2
